+++ screens/SendConfirmScreen.js
@@ -16,8 +16,6 @@ import BigNumber from "bignumber.js";
 import Swipeable from "react-native-swipeable";
 import Ionicons from "react-native-vector-icons/Ionicons";
 
-import SLPSDK from "slp-sdk";
-
 import { Button, T, H1, H2, Spacer } from "../atoms";
 
 import { type TokenData } from "../data/tokens/reducer";
@@ -44,7 +42,7 @@ import {
 import { utxosByAccountSelector } from "../data/utxos/selectors";
 import { spotPricesSelector, currencySelector } from "../data/prices/selectors";
 
-const SLP = new SLPSDK();
+import { SLP } from "../utils/slp-sdk-utils";
 
 const SWIPEABLE_WIDTH_PERCENT = 78;
 
@@ -70,6 +68,8 @@ const SwipeButtonContainer = styled(View)`
   border-radius: 32px;
   width: ${SWIPEABLE_WIDTH_PERCENT}%;
   align-self: center;
+  border-width: 2px;
+  border-color: ${props => props.theme.primary700};
 `;
 
 const ButtonsContainer = styled(View)`
@@ -90,6 +90,7 @@ const SwipeContent = styled(View)`
   padding-right: 10px;
   align-items: flex-end;
   justify-content: center;
+
   background-color: ${props =>
     props.activated ? props.theme.success500 : props.theme.pending500};
 `;
@@ -320,7 +321,7 @@ const SendConfirmScreen = ({
                 leftActionActivationDistance={
                   Dimensions.get("window").width *
                   (SWIPEABLE_WIDTH_PERCENT / 100) *
-                  0.8
+                  0.7
                 }
                 leftContent={
                   <SwipeContent activated={confirmSwipeActivated}>