+++ utils/balance-utils.js
@@ -1,7 +1,6 @@
 // @flow
 
 import BigNumber from "bignumber.js";
-import SLPSDK from "slp-sdk";
 
 import {
   currencyDecimalMap,
@@ -10,10 +9,11 @@ import {
 } from "./currency-utils";
 import { type Balances } from "../data/selectors";
 
-const SLP = new SLPSDK();
+import { SLP } from "./slp-sdk-utils";
 
 const getHistoricalBchTransactions = async (
   address: string,
+  addressSlp: string,
   latestBlock: number
 ) => {
   if (!address) {
@@ -30,6 +30,12 @@ const getHistoricalBchTransactions = async (
             },
             {
               "out.e.a": address.slice(12)
+            },
+            {
+              "in.e.a": addressSlp.slice(12)
+            },
+            {
+              "out.e.a": addressSlp.slice(12)
             }
           ],
           "out.h1": {
@@ -65,14 +71,14 @@ const getHistoricalBchTransactions = async (
   const result = await request.json();
 
   // combine confirmed and unconfirmed
-  const transactions = [...result.c, ...result.u];
+  const transactions = result.errors ? [] : [...result.c, ...result.u];
 
   return transactions;
 };
 
 const getHistoricalSlpTransactions = async (
   address: string,
-  slpAddress: string,
+  addressSlp: string,
   latestBlock: number
 ) => {
   if (!address) return [];
@@ -90,10 +96,10 @@ const getHistoricalSlpTransactions = async (
               "slp.detail.outputs.address": SLP.Address.toSLPAddress(address)
             },
             {
-              "in.e.a": SLP.Address.toSLPAddress(slpAddress)
+              "in.e.a": SLP.Address.toSLPAddress(addressSlp)
             },
             {
-              "slp.detail.outputs.address": SLP.Address.toSLPAddress(slpAddress)
+              "slp.detail.outputs.address": SLP.Address.toSLPAddress(addressSlp)
             }
           ],
           "slp.valid": true,